TASK: Improve Pure Data
Try make more UI elements in Pure Data but
note that there's a lot of copy/paste so
a lot of the ctlout CC numbers are duplicated

That's actually caused a rethink. I think there
only ever will be 8 CCs (probably 100 to 107)
then there'll be a concept of "active screen"
and how 100-107 are used will differ according to
the sub-part being edited at the time.