TASK: change display updating and loop()
Previously sim was repainting (InavlidateRect) for
every object drawn. Also it was not calling loop()
in Display.cpp

Set timer to 5ms to regularly call loop() and on
same timer force a WM_PAINT to repaint.

Still invalidating "too much" and causing display
flickers so need to determine "what changed" and
only repaint that area.