TASK: Finish implementation of most of Arpeggiator
Crikey this is a good one - the Arpeggiator is
my new favourite toy. The one control not "wired
up" yet is the "Octave" though I'm not sure it's
even needed but the plan was to extend the sequence
by N octaves by playing the note array and then the
same +12 and so on for however many octaves are
selected. But I'm not sure if it really needs that.