TASK: Bring INO, Novation, Ctrlr into line
Several fixeds to INO and ensure it reponds to
the same control ranges as put out by both
Ctrlr and the Novation Remote.

As Novation now updated to OS 2.3 so it can
do steps greater than unity revert to previous
scheme of using "ranges" for CC selectors so
they could be driven by normal pots

Update pictures to match definitions.