TASK: Mainly to do with wavetable samples
Most of this change is about finally getting
a table of a few Wavetable samples (Piano,
Violin, Flute, Guitar, etc) to work and that
includes adding a selector to Ctrlr and also
adding a rotary CC to change the wavetable
volume in the mix.

Then update things like the Novation panel
definition and the Ctrlr image to match (it has
the reassigned CC numbers)

Also avoid using CC 100/101 as these are really
for RPN and should be avoided.