TASK: Loads more voices and other changes
Add a coniderable number more of the
instrument samples and make sure the INO
and the Ctrlr allow them to be selected.

Change refs to lfo "amplitude" to be depth.

Change depth so it has far less effect so
the range of the control can be used more

Change LFO frequency to be in two parts
with the first half selecting 0..20Hz then
the next selects 20..200Hz.