TASK: Probably the last of the samples
Build is now up to 94% so this is very likely
the last samples that can fit in. There's a few
more that could be added but it as they are
"big ones" that are left it would likely mean
cutting a few unless something intelligent
can be done where instruments actually share
samples - or perhaps a few "lo-fi" samples
from smaller SF2 files?