TASK: Add a sequence recorder
Main change is the addition of a working
sequence recorder.

In doing so switch to using elapsedMillis
variables (to save time subtractions)

Also fix issues with Arpeggiator defaulting
to wrong ocatve and other issues

Also use a new synth patch which moves the
ADSR back up the chain with a view to adding
some paraphony to it with multiple copies
of the sound generation blocks.