TASK: Make (non USB) MIDI callbacks work
Trick to making MIDI callbacks work is simply
doing a MIDI.read() in loop)()

So then remove the polling MIDI code that is
no longer required.

Also remove a lot of dead "test" code (it's
always in git anyway!)

Fix an if (xxx=) that should have been if (xxx==).

Note to self: always take pictures of wiring
so you can later reinstate if you take it apart.