TASK: Really begin to move to Poly synth
MewSynth.sip was just a step on the way to
making the synth class based and poly but there
were things that could not be achieved with the
Manicken audio design tool at that time.

But now it looks achievable so rename to "PhatBass"
as that's closer to the desired solution and add
(for now) both the .ZIP saved straight from the
design tool then, to allow for version tracking, an
unpacked copy of same in the PhatBass sub-dir.